+++ CHANGELOG.md
@@ -19,7 +19,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - `DateFunctions.now()` returns the current time in UTC as defined by the GREL funtions (not in local timezone).
 - `StringFunctions.replace()` now performs string replace as well as regex replace.
-- `StringFunctions.match()` fixed GREL regex matching
+- `StringFunctions.match()`: fixed GREL regex matching
+- `StringFunctions.toNumber()`: expanded conversion to `Number`, basically covering all numbers, not just integers.
 
 ## [0.8.2] - 2022-09-26